to typescript + center loading spinner
